Add in get/setReadOnly methods Also, refactoring to so we're not just in one big file for everything.
