+++ sdk/appconfiguration/app-configuration/test/index.spec.ts
@@ -8,10 +8,8 @@ import {
   toSortedArray,
   assertEqualSettings
 } from "./testHelpers";
-import * as dotenv from "dotenv";
 import { AppConfigurationClient } from "../src";
-
-dotenv.config();
+import { formatETagForMatchHeaders, formatWildcards, extractAfterTokenFromNextLink } from '../src/internal/helpers';
 
 describe("AppConfigurationClient", () => {
   const settings: Array<{ key: string; label?: string }> = [];
@@ -922,14 +920,14 @@ describe("AppConfigurationClient", () => {
     it("undefined", () => {
       assert.equal(
         undefined,
-        AppConfigurationClient["formatETagForMatchHeaders"]({
+        formatETagForMatchHeaders({
           etag: undefined
         })
       );
 
       assert.equal(
         '"etagishere"',
-        AppConfigurationClient["formatETagForMatchHeaders"]({
+        formatETagForMatchHeaders({
           etag: "etagishere"
         })
       );
@@ -938,7 +936,7 @@ describe("AppConfigurationClient", () => {
 
   describe("formatWildcards", () => {
     it("undefined", () => {
-      const result = AppConfigurationClient["formatWildcards"]({
+      const result = formatWildcards({
         keys: undefined,
         labels: undefined
       });
@@ -948,7 +946,7 @@ describe("AppConfigurationClient", () => {
     });
 
     it("single values only", () => {
-      const result = AppConfigurationClient["formatWildcards"]({
+      const result = formatWildcards({
         keys: ["key1"],
         labels: ["label1"]
       });
@@ -958,7 +956,7 @@ describe("AppConfigurationClient", () => {
     });
 
     it("multiple values", () => {
-      const result = AppConfigurationClient["formatWildcards"]({
+      const result = formatWildcards({
         keys: ["key1", "key2"],
         labels: ["label1", "label2"]
       });
@@ -970,7 +968,7 @@ describe("AppConfigurationClient", () => {
 
   describe("extractAfterTokenFromNextLink", () => {
     it("token is extracted and properly unescaped", () => {
-      let token = AppConfigurationClient["extractAfterTokenFromNextLink"](
+      let token = extractAfterTokenFromNextLink(
         "/kv?key=someKey&api-version=1.0&after=bGlah%3D"
       );
       assert.equal("bGlah=", token);
