+++ sdk/appconfiguration/app-configuration/test/testHelpers.ts
@@ -9,6 +9,7 @@ import * as assert from "assert";
 // allow loading from a .env file as an alternative to defining the variable
 // in the environment
 import * as dotenv from "dotenv";
+import { RestError } from '@azure/core-http';
 dotenv.config();
 
 export function getConnectionStringFromEnvironment(): string {
@@ -29,6 +30,10 @@ export async function deleteKeyCompletely(keys: string[], client: AppConfigurati
   });
 
   for await (const setting of settingsIterator) {
+    if (setting.locked) {
+      await client.clearReadOnly(setting);
+    }
+
     await client.deleteConfigurationSetting(setting.key, { label: setting.label });
   }
 }
@@ -69,4 +74,18 @@ export function assertEqualSettings(
     return { key: setting.key, label: setting.label, value: setting.value };
   });
   assert.deepEqual(expected, actual);
+}
+
+export async function assertThrowsRestError(testFunction: () => Promise<any>, expectedStatusCode: number, message: string) : Promise<void> {
+  try {
+    await testFunction();
+    assert.fail("No error thrown");
+  } catch (err) {
+    if (err instanceof RestError) {
+      assert.equal(expectedStatusCode, err.statusCode);
+      return;
+    }
+
+    assert.fail(`${message}: Caught error: ${err}`);
+  }
 }
